wip: updating upload handler code
