+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/upload/certifiedProduct/CertifiedProductHandler2015Version3.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.apache.commons.csv.CSVRecord;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
@@ -16,11 +17,12 @@ import gov.healthit.chpl.entity.listing.pending.PendingCertificationResultTestPr
 import gov.healthit.chpl.entity.listing.pending.PendingCertifiedProductEntity;
 import gov.healthit.chpl.upload.certifiedProduct.template.TemplateColumnIndexMap;
 import gov.healthit.chpl.upload.certifiedProduct.template.TemplateColumnIndexMap2015Version3;
+import gov.healthit.chpl.util.ErrorMessageUtil;
 
 /**
  * Adds ICS Source (family), Removes columns G1 and G2 for 170.315(g)(7),
  * Adds Test Tool + Test Tool Version + Test Data Version, Test Data Alteration, Test Data Alteration Description for criteria b8,
- * Removes Test Tool + Test Tool Version + Test Data fields for criteria f5
+ * Removes Test Tool + Test Tool Version + Test Data fields for criteria f5.
  * @author kekey
  *
  */
@@ -30,7 +32,9 @@ public class CertifiedProductHandler2015Version3 extends CertifiedProductHandler
     private static final Logger LOGGER = LogManager.getLogger(CertifiedProductHandler2015Version3.class);
     private TemplateColumnIndexMap templateColumnIndexMap;
 
-    public CertifiedProductHandler2015Version3() {
+    @Autowired
+    public CertifiedProductHandler2015Version3(final ErrorMessageUtil msgUtil) {
+        super(msgUtil);
         templateColumnIndexMap = new TemplateColumnIndexMap2015Version3();
     }
